requirements.txt
rstr
plotly
numpy
scipy
matplotlib
scikit_learn
sklearn
gitpython